+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## Next Release
+
+- Remove `carbon_offset` (`CarbonOffset`) parameter from shipment creation and purchase flows
+  - This parameter is no longer supported by the EasyPost API
+- Remove `CarbonOffset` model
+
 ## v5.8.0 (2023-10-11)
 
 - Add `RetrieveApiKeysForUser` to `ApiKey` service